alert mac user how to proper launch the installer
If the user launch the installer when opening the triangles (instead of double-clicking on the folder), Finder can't find the lib/riv.jar and the installation fails.
Would this happen, this commit shows the following message
To install RuralInvest, please double-click on the folder that contains the installer, then double-click again on the Installer app.
Please see the Quickstart for more information

git-svn-id: https://svn.fao.org/projects/ruralinvest/RIV4/trunk@67564 d7fa552c-c32d-11e1-9f2d-6d39f2a56844